+++ src/Controller/Info/AdminRosterController.php
@@ -4,7 +4,6 @@ namespace App\Controller\Info;
 
 use App\Controller\Controller;
 use App\Domain\Admin\Repository\AdminRepository;
-use App\Enum\AdminRanks;
 use App\Enum\PermissionsFlags;
 use Psr\Http\Message\ResponseInterface;
 use DI\Attribute\Inject;